oh wow I didn't commit anything yet
Added 5 functions to filter images
Added a caseiterable enum with a String raw value to track these functions.

Added a View Controller and storyboard views in the IB to provide a user interface. This is composed of the selected image view (Will incorporate ImagePickerDelegate), an empty view that will switch on a series of custom controls using the final part, the PickerView which will select which filter to use and display the correct set of controls in the middle window.

Incorporated uipickerviewdelegate and datasource

added outlines for 5 filter controls